test/touchevents/single-touch.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Touch Events: Single-touch tests</title>
  <meta name="viewport" content="width=device-width">
  <script src="../testharness.js"></script>
  <script>
    setup({explicit_done: true});

    function run() {
      var target = document.getElementById("target");

      var test_click = async_test("Interaction with mouse events");
      var test_touchstart = async_test("touchstart event received");
      var test_touchend = async_test("touchend event received");

      var touchstart_received = false;

      on_event(target, "touchstart", function onTouchStart(ev) {
        touchstart_received = true;
        test_touchstart.done();

        test(function() {
          assert_true(ev.target instanceof Element, "target must be an Element.");
          assert_equals(ev.relatedTarget, null, "relatedTarget must be null for touchstart event.");
        }, "touchstart TouchEvent target attributes are correct.");

        test(function() {
          var attrs = ["altKey", "metaKey", "ctrlKey", "shiftKey"];
          for (var i = 0; i < attrs.length; i++) {
            assert_true(attrs[i] in ev);
          }
        }, "touchstart TouchEvent modifier key attributes are present.");

        test(function() {
          assert_equals(ev.touches.length, 1, "One touch point.");
          assert_equals(ev.changedTouches.length, 1, "One changed touch point.");
          assert_equals(ev.targetTouches.length, 1, "One target touch point.");
        }, "touchstart TouchList lengths are correct.");

        var t = ev.touches[0];
        var ct = ev.changedTouches[0];
        var tt = ev.targetTouches[0];

        test(function() {
          assert_true(ev instanceof TouchEvent, "Event is a TouchEvent.");
          assert_true(ev.touches instanceof TouchList, "touches is a TouchList");
          assert_true(t instanceof Touch, "touches[0] is a Touch");
        }, "Interface names are correct.");

        test(function() {
          assert_equals(t.identifier, 0, "touches identifier is 0.");
          assert_equals(ct.identifier, 0, "changedTouches identifier is 0.");
          assert_equals(tt.identifier, 0, "targetTouches identifier is 0.");
        }, "Touch identifiers are correct.");

        test(function() {
          assert_true(t.rotationAngle >= 0, "rotationAngle must be >= 0.");
          assert_true(t.rotationAngle < 90, "rotationAngle must be > 90.");

          assert_true(t.radiusX > 0, "radiusX is positive.");
          assert_true(t.radiusY > 0, "radiusY is positive.");

          assert_true(t.force >= 0, "force >= 0.");
          assert_true(t.force <= 1, "force <= 1.");
        }, "Touch attribute values are valid");

        test(function() {
          var attrs = ["screenX", "screenY", "clientX", "clientY",
                       "pageX", "pageY"];
          for (var i = 0; i < attrs.length; i++) {
            assert_true(attrs[i] in t);
          }
        }, "Touch location attributes are present.");

      });

      on_event(target, "touchend", function onTouchStart(ev) {
        test_touchend.step(function() {
          assert_true(touchstart_received, "touchend follows touchstart");
        });
        test_touchend.done();

        test(function() {
          var attrs = ["altKey", "metaKey", "ctrlKey", "shiftKey"];
          for (var i = 0; i < attrs.length; i++) {
            assert_true(attrs[i] in ev);
          }
        }, "touchend TouchEvent modifier key attributes are present.");

        test(function() {
          assert_equals(ev.touches.length, 0, "Zero touch points.");
          assert_equals(ev.changedTouches.length, 1, "One changed touch point.");
          assert_equals(ev.targetTouches.length, 0, "Zero target touch points.");
        }, "touchend TouchList lengths are correct.");

        var t = ev.changedTouches[0];

        test(function() {
          assert_equals(t.identifier, 0, "changedTouches identifier is 0.");
        }, "Touch identifier is correct.");

        test(function() {
          assert_true(t.rotationAngle >= 0, "rotationAngle must be >= 0.");
          assert_true(t.rotationAngle < 90, "rotationAngle must be > 90.");

          assert_true(t.radiusX > 0, "radiusX is positive.");
          assert_true(t.radiusY > 0, "radiusY is positive.");

          assert_true(t.force >= 0, "force >= 0.");
          assert_true(t.force <= 1, "force <= 1.");
        }, "Touch attribute values are valid");

        test(function() {
          var attrs = ["screenX", "screenY", "clientX", "clientY",
                       "pageX", "pageY"];
          for (var i = 0; i < attrs.length; i++) {
            assert_true(attrs[i] in t);
          }
        }, "Touch location attributes are present.");

        done();
      });


      on_event(target, "click", function onClick(ev) {
        test_click.step(function() {
          assert_true(touchstart_received,
            "The touchstart event must be dispatched before any mouse " +
            "events. (If this fails, it might mean that the user agent does " +
            "not implement W3C touch events at all.)"
          );
        });
        test_click.done();

        if (!touchstart_received) {
          // Abort the tests.  If touch events are not supported, then most of
          // the other event handlers will never be called, and the test will
          // time out with misleading results.
          done();
        }
      });
    }
  </script>
  <style>
    #target {
      background: yellow;
      border: 1px solid orange;
      padding: 2em;
    }
  </style>
</head>
<body onload="run()">
  <h1>Touch area tests</h1>
  <p id="target">
    Tap this element with one finger (or other pointing device).
  </p>
  <div id="log"></div>
</body>
</html>
